+++ elasticsearch-singledocument/src/test/java/org/vertexium/elasticsearch/ElasticsearchSingleDocumentSearchIndexTest.java
@@ -1,5 +1,14 @@
 package org.vertexium.elasticsearch;
 
+import org.junit.ClassRule;
+
 public class ElasticsearchSingleDocumentSearchIndexTest extends ElasticsearchSingleDocumentSearchIndexTestBase {
 
+    @ClassRule
+    public static ElasticsearchResource elasticsearchResource = new ElasticsearchResource();
+
+    @Override
+    protected ElasticsearchResource getElasticsearchResource() {
+        return elasticsearchResource;
+    }
 }