+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added: ability to store extended data rows on an element
 * Added: Elasticsearch 2.x support
+* Fixed: Issue #135. Passing FetchHint.NONE when retrieving vertices from Accumulo using Elasticsearch will now properly return the vertices rather than an empty Iterable
+* Deprecated: EdgeCountScoringStrategy
 
 # v2.5.4